chore: separate Video component
