--- client/src/components/Classroom/CamButton/CamButton.css
@@ -1,13 +1,0 @@
-.cam-input-btn {
-  background-color: rgba(25, 216, 223, 0.9);
-  border: 1px solid rgba(0,0,0,0.9);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 2em;
-  height: 2em;
-  border-radius: 50%;
-  cursor: pointer;
-  padding: 0.5em;
-  margin: 0.5em;
-}
