+++ client/src/components/Classroom/MicButton/MicButton.css
@@ -1,13 +1,1 @@
-.mic-input-btn {
-  background-color: rgba(223, 141, 25, 0.65);
-  border: 1px solid rgba(0,0,0,0.9);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 2em;
-  height: 2em;
-  border-radius: 50%;
-  cursor: pointer;
-  padding: 0.5em;
-  margin: 0.5em;
-}
+
