--- client/src/components/Classroom/PhoneButton/PhoneButton.css
@@ -1,13 +1,0 @@
-.phone-input-btn {
-  background-color: rgba(249, 52, 52, 0.738);
-  border: 1px solid rgba(0, 0, 0, 0.9);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 2em;
-  height: 2em;
-  border-radius: 50%;
-  cursor: pointer;
-  padding: 0.5em;
-  margin: 0.5em;
-}
