+++ client/src/components/pages/ClassroomPage.css
@@ -27,6 +27,8 @@
   border-radius: 0.3em;
 }
 
+
+
 .peer-video {
   padding-top: 1em;
   background-color:#060606;
@@ -54,20 +56,26 @@
   border-radius: 50%;
   cursor: pointer;
   padding: 0.5em;
+}
+
+.video-controls {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  padding: 0.5em;
   margin: 0.5em;
 }
 
 .cam-input-btn {
-  background-color: rgba(190, 234, 235, 0.9);
-  
+  background-color: rgba(27, 230, 234, 0.9) !important; 
 }
 
 .mic-input-btn {
-  background-color: rgba(237, 235, 233, 0.65);
+  background-color: rgba(237, 235, 233, 0.65) !important;
 }
 
 .phone-input-btn {
-  background-color: rgba(249, 52, 52, 0.738);
+  background-color: rgba(249, 52, 52, 0.738) !important;
 }
 
 @media screen and (max-width: 390px) {
@@ -158,12 +166,13 @@
   }
   
   .my-video {
-    padding-top: 0.2em;
+    padding-top: 0.1em;
     background-color:#060606;
+    padding-bottom: 0.5em;
   }
 
   .peer-video {
-    padding-top: 1em;
+    padding-top: 1.5em;
     background-color:#060606;
     border-radius: 0.3em;
   }
